sqlcoder_cpu_gpu/requirements.txt
sqlparse==0.4.4
runpod==1.6.2
llama-cpp-python==0.2.39